Raymarcher now renders into a pixel buffer instead of directly to an image file

+++ src/vec.cpp
@@ -1,6 +1,6 @@
 #include "vec.h"
 
-Vec3::Vec3(float x, float y, float z) {
+Vec3::Vec3(double x, double y, double z) {
 	this->x = x;
 	this->y = y;
 	this->z = z;
